# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  subjectId?: number | null;
  priority: TaskPriority;
  dueDate?: string | null;
  completed: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  subjectId?: number | null;
  priority: CreateTaskBodyPriority;
  dueDate?: string | null;
  completed?: boolean;
}

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  subjectId?: number | null;
  priority?: UpdateTaskBodyPriority;
  dueDate?: string | null;
  completed?: boolean;
}

export interface Subject {
  id: number;
  name: string;
  color: string;
  icon?: string | null;
  createdAt: string;
}

export interface CreateSubjectBody {
  name: string;
  color: string;
  icon?: string | null;
}

export interface UpdateSubjectBody {
  name?: string;
  color?: string;
  icon?: string | null;
}

export type StudySessionSessionType =
  (typeof StudySessionSessionType)[keyof typeof StudySessionSessionType];

export const StudySessionSessionType = {
  pomodoro: "pomodoro",
  regular: "regular",
  exam_prep: "exam_prep",
} as const;

export interface StudySession {
  id: number;
  subjectId?: number | null;
  date: string;
  startTime: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  sessionType: StudySessionSessionType;
  notes?: string | null;
  createdAt: string;
}

export type CreateSessionBodySessionType =
  (typeof CreateSessionBodySessionType)[keyof typeof CreateSessionBodySessionType];

export const CreateSessionBodySessionType = {
  pomodoro: "pomodoro",
  regular: "regular",
  exam_prep: "exam_prep",
} as const;

export interface CreateSessionBody {
  subjectId?: number | null;
  date: string;
  startTime: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  sessionType: CreateSessionBodySessionType;
  notes?: string | null;
}

export type UpdateSessionBodySessionType =
  (typeof UpdateSessionBodySessionType)[keyof typeof UpdateSessionBodySessionType];

export const UpdateSessionBodySessionType = {
  pomodoro: "pomodoro",
  regular: "regular",
  exam_prep: "exam_prep",
} as const;

export interface UpdateSessionBody {
  subjectId?: number | null;
  date?: string;
  startTime?: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  sessionType?: UpdateSessionBodySessionType;
  notes?: string | null;
}

export interface StatsSummary {
  totalTasks: number;
  completedTasks: number;
  pendingTasks: number;
  overdueCount: number;
  totalStudyMinutes: number;
  weekStudyMinutes: number;
  todayStudyMinutes: number;
  completionRate: number;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  lastStudyDate?: string | null;
  studiedToday: boolean;
}

export interface SubjectStats {
  subjectId: number;
  subjectName: string;
  subjectColor: string;
  totalTasks: number;
  completedTasks: number;
  completionRate: number;
  totalStudyMinutes: number;
}

export type ListTasksParams = {
  subjectId?: number;
  priority?: ListTasksPriority;
  completed?: boolean;
};

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type ListSessionsParams = {
  subjectId?: number;
  date?: string;
};