# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Study Planner API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management
  - name: subjects
    description: Subject management
  - name: sessions
    description: Study session management
  - name: stats
    description: Statistics and progress
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: completed
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a new subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"

  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubjectBody"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List study sessions
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"

  /sessions/{id}:
    patch:
      operationId: updateSession
      tags: [sessions]
      summary: Update a study session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a study session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get overall progress summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/streak:
    get:
      operationId: getStreak
      tags: [stats]
      summary: Get current study streak
      responses:
        "200":
          description: Current streak info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakInfo"

  /stats/by-subject:
    get:
      operationId: getStatsBySubject
      tags: [stats]
      summary: Get task completion stats grouped by subject
      responses:
        "200":
          description: Per-subject stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        subjectId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date
          nullable: true
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - priority
        - completed
        - createdAt
        - updatedAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        subjectId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date
          nullable: true
        completed:
          type: boolean
      required:
        - title
        - priority

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        subjectId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date
          nullable: true
        completed:
          type: boolean

    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        icon:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdAt

    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        icon:
          type: string
          nullable: true
      required:
        - name
        - color

    UpdateSubjectBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        icon:
          type: string
          nullable: true

    StudySession:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        sessionType:
          type: string
          enum: [pomodoro, regular, exam_prep]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - startTime
        - sessionType
        - createdAt

    CreateSessionBody:
      type: object
      properties:
        subjectId:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        sessionType:
          type: string
          enum: [pomodoro, regular, exam_prep]
        notes:
          type: string
          nullable: true
      required:
        - date
        - startTime
        - sessionType

    UpdateSessionBody:
      type: object
      properties:
        subjectId:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        sessionType:
          type: string
          enum: [pomodoro, regular, exam_prep]
        notes:
          type: string
          nullable: true

    StatsSummary:
      type: object
      properties:
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        pendingTasks:
          type: integer
        overdueCount:
          type: integer
        totalStudyMinutes:
          type: integer
        weekStudyMinutes:
          type: integer
        todayStudyMinutes:
          type: integer
        completionRate:
          type: number
      required:
        - totalTasks
        - completedTasks
        - pendingTasks
        - overdueCount
        - totalStudyMinutes
        - weekStudyMinutes
        - todayStudyMinutes
        - completionRate

    StreakInfo:
      type: object
      properties:
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        lastStudyDate:
          type: string
          format: date
          nullable: true
        studiedToday:
          type: boolean
      required:
        - currentStreak
        - longestStreak
        - studiedToday

    SubjectStats:
      type: object
      properties:
        subjectId:
          type: integer
        subjectName:
          type: string
        subjectColor:
          type: string
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        completionRate:
          type: number
        totalStudyMinutes:
          type: integer
      required:
        - subjectId
        - subjectName
        - subjectColor
        - totalTasks
        - completedTasks
        - completionRate
        - totalStudyMinutes