# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all tasks
 */
export const ListTasksQueryParams = zod.object({
  subjectId: zod.coerce.number().optional(),
  priority: zod.enum(["low", "medium", "high"]).optional(),
  completed: zod.coerce.boolean().optional(),
});

export const ListTasksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  subjectId: zod.number().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  dueDate: zod.coerce.date().nullish(),
  completed: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListTasksResponse = zod.array(ListTasksResponseItem);

/**
 * @summary Create a new task
 */
export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  subjectId: zod.number().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  dueDate: zod.coerce.date().nullish(),
  completed: zod.boolean().optional(),
});

/**
 * @summary Get a task by ID
 */
export const GetTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  subjectId: zod.number().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  dueDate: zod.coerce.date().nullish(),
  completed: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
  subjectId: zod.number().nullish(),
  priority: zod.enum(["low", "medium", "high"]).optional(),
  dueDate: zod.coerce.date().nullish(),
  completed: zod.boolean().optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  subjectId: zod.number().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  dueDate: zod.coerce.date().nullish(),
  completed: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all subjects
 */
export const ListSubjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  color: zod.string(),
  icon: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListSubjectsResponse = zod.array(ListSubjectsResponseItem);

/**
 * @summary Create a new subject
 */
export const CreateSubjectBody = zod.object({
  name: zod.string(),
  color: zod.string(),
  icon: zod.string().nullish(),
});

/**
 * @summary Update a subject
 */
export const UpdateSubjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSubjectBody = zod.object({
  name: zod.string().optional(),
  color: zod.string().optional(),
  icon: zod.string().nullish(),
});

export const UpdateSubjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  color: zod.string(),
  icon: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a subject
 */
export const DeleteSubjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List study sessions
 */
export const ListSessionsQueryParams = zod.object({
  subjectId: zod.coerce.number().optional(),
  date: zod.date().optional(),
});

export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  subjectId: zod.number().nullish(),
  date: zod.coerce.date(),
  startTime: zod.string(),
  endTime: zod.string().nullish(),
  durationMinutes: zod.number().nullish(),
  sessionType: zod.enum(["pomodoro", "regular", "exam_prep"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Create a study session
 */
export const CreateSessionBody = zod.object({
  subjectId: zod.number().nullish(),
  date: zod.coerce.date(),
  startTime: zod.string(),
  endTime: zod.string().nullish(),
  durationMinutes: zod.number().nullish(),
  sessionType: zod.enum(["pomodoro", "regular", "exam_prep"]),
  notes: zod.string().nullish(),
});

/**
 * @summary Update a study session
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSessionBody = zod.object({
  subjectId: zod.number().nullish(),
  date: zod.coerce.date().optional(),
  startTime: zod.string().optional(),
  endTime: zod.string().nullish(),
  durationMinutes: zod.number().nullish(),
  sessionType: zod.enum(["pomodoro", "regular", "exam_prep"]).optional(),
  notes: zod.string().nullish(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.number(),
  subjectId: zod.number().nullish(),
  date: zod.coerce.date(),
  startTime: zod.string(),
  endTime: zod.string().nullish(),
  durationMinutes: zod.number().nullish(),
  sessionType: zod.enum(["pomodoro", "regular", "exam_prep"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a study session
 */
export const DeleteSessionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get overall progress summary
 */
export const GetStatsSummaryResponse = zod.object({
  totalTasks: zod.number(),
  completedTasks: zod.number(),
  pendingTasks: zod.number(),
  overdueCount: zod.number(),
  totalStudyMinutes: zod.number(),
  weekStudyMinutes: zod.number(),
  todayStudyMinutes: zod.number(),
  completionRate: zod.number(),
});

/**
 * @summary Get current study streak
 */
export const GetStreakResponse = zod.object({
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  lastStudyDate: zod.coerce.date().nullish(),
  studiedToday: zod.boolean(),
});

/**
 * @summary Get task completion stats grouped by subject
 */
export const GetStatsBySubjectResponseItem = zod.object({
  subjectId: zod.number(),
  subjectName: zod.string(),
  subjectColor: zod.string(),
  totalTasks: zod.number(),
  completedTasks: zod.number(),
  completionRate: zod.number(),
  totalStudyMinutes: zod.number(),
});
export const GetStatsBySubjectResponse = zod.array(
  GetStatsBySubjectResponseItem,
);