# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalTasks: number;
  completedTasks: number;
  pendingTasks: number;
  overdueCount: number;
  totalStudyMinutes: number;
  weekStudyMinutes: number;
  todayStudyMinutes: number;
  completionRate: number;
}