# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudySessionSessionType } from "./studySessionSessionType";

export interface StudySession {
  id: number;
  subjectId?: number | null;
  date: Date;
  startTime: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  sessionType: StudySessionSessionType;
  notes?: string | null;
  createdAt: Date;
}