# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  subjectId?: number | null;
  priority: TaskPriority;
  dueDate?: Date | null;
  completed: boolean;
  createdAt: Date;
  updatedAt: Date;
}