# lib/api-zod/src/generated/types/updateSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSessionBodySessionType } from "./updateSessionBodySessionType";

export interface UpdateSessionBody {
  subjectId?: number | null;
  date?: Date;
  startTime?: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  sessionType?: UpdateSessionBodySessionType;
  notes?: string | null;
}