# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  subjectId?: number | null;
  priority?: UpdateTaskBodyPriority;
  dueDate?: Date | null;
  completed?: boolean;
}